+++ plugins/catalog-import/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-import
 
+## 0.9.8-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.1-next.1
+  - @backstage/integration-react@1.1.13-next.2
+  - @backstage/plugin-catalog-react@1.6.0-next.2
+  - @backstage/config@1.0.7
+  - @backstage/core-plugin-api@1.5.1
+
 ## 0.9.8-next.1
 
 ### Patch Changes